python_service/requirements.txt
google-generativeai
python-dotenv

Flask==3.0.0
pyswisseph==2.10.3.2
